[TASK] code improvements
Relates: #85
